+++ utils/config.js
@@ -1,10 +1,11 @@
-require('dotenv').config()
+import dotenv from 'dotenv';
+dotenv.config();
 
 const PORT = process.env.PORT
 const MONGODB_URI = process.env.NODE_ENV === 'test'
 	? process.env.MONGODB_URI_TEST
 	: process.env.MONGODB_URI
 
-module.exports = {
+export default {
 	PORT, MONGODB_URI
 }
