+++ utils/logger.js
@@ -7,6 +7,6 @@ const error = (... params) => {
 	console.error(params)
 }
 
-module.exports = {
+export default {
 	info, error
 }
