+++ utils/middleware.js
@@ -1,5 +1,4 @@
-const logger = require('./logger')
-// const jwt = require('jsonwebtoken')
+import logger from './logger.js'
 
 const requestLogger = (request, response, next) => {
 	logger.info('Method:', request.method)
@@ -26,4 +25,4 @@ const errorHandler = (error, request, response, next) => {
 	next(error)
 }
 
-module.exports = {requestLogger, unknownEndpoint, errorHandler}
+export default {requestLogger, unknownEndpoint, errorHandler}
